feat(Pointer): add variable downward length on bezier pointer
The Bezier Pointer Renderer now can have the downward length of the
beam set as well as setting the forward length of the beam.

This is now done by making the `maximumLength` parameter a Vector2
instead of a float where the `x` value represents the forward beam
length and the `y` value represents the downward beam length.

An issue with the document generator has also been fixed causing
more documentation to be generated than jsut for the above feature.